Minor refactoring.

+++ src/java/org/orbeon/oxf/pipeline/InitUtils.java
@@ -1,15 +1,15 @@
 /**
- *  Copyright (C) 2004 Orbeon, Inc.
+ * Copyright (C) 2009 Orbeon, Inc.
  *
- *  This program is free software; you can redistribute it and/or modify it under the terms of the
- *  GNU Lesser General Public License as published by the Free Software Foundation; either version
- *  2.1 of the License, or (at your option) any later version.
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU Lesser General Public License as published by the Free Software Foundation; either version
+ * 2.1 of the License, or (at your option) any later version.
  *
- *  This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- *  See the GNU Lesser General Public License for more details.
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
  *
- *  The full text of the license is available at http://www.gnu.org/copyleft/lesser.html
+ * The full text of the license is available at http://www.gnu.org/copyleft/lesser.html
  */
 package org.orbeon.oxf.pipeline;
 
@@ -82,7 +82,7 @@ public class InitUtils {
             // Set cache size
             final Integer cacheMaxSize = Properties.instance().getPropertySet().getInteger(CACHE_SIZE_PROPERTY);
             if (cacheMaxSize != null)
-                ObjectCache.instance().setMaxSize(pipelineContext, cacheMaxSize.intValue());
+                ObjectCache.instance().setMaxSize(pipelineContext, cacheMaxSize);
 
             // Start execution
             processor.reset(pipelineContext);
@@ -210,6 +210,7 @@ public class InitUtils {
      *
      * @param servletContext            required ServletContext instance
      * @param session                   optional HttpSession object
+     * @param localMap
      * @param logger                    required logger
      * @param logMessagePrefix          required prefix for log messages
      * @param message                   optional message to display whether there is a processor to run or not
